call_center_analytics/requirements.txt
langgraph==0.2.45
langchain==0.3.9
langchain-openai==0.2.9
langchain-community==0.3.8
streamlit==1.40.1
openai==1.57.2
openai-whisper==20250625
python-dotenv==1.0.1
pydantic==2.10.3
typing-extensions==4.12.2
pandas==2.2.0
